+++ llm_infer/client/__init__.py
@@ -80,6 +80,7 @@ from .errors import (
     BackendTimeoutError,
     BackendUnavailableError,
     ConfigError,
+    FallbackAmbiguityError,
     ModelConflictError,
 )
 from .factory import Factory
@@ -170,6 +171,7 @@ __all__ = [
     "BackendTimeoutError",
     "BackendUnavailableError",
     "ConfigError",
+    "FallbackAmbiguityError",
     "ModelConflictError",
 ]